auth: switch back to launchWebAuthFlow
Using launchWebAuthFlow is (AFAICT) the only way to do auth noninteractively, which is a requirement for automatic backups*.
We can't use XHR/fetch to resolve the redirect because CORS and credentials aren't compatible (and you can't read the hash out of redirect URLs, apparently).
We can't use an iframe, because [the spec](https://tools.ietf.org/html/draft-ietf-oauth-v2-23#section-10.13) and real implementations forbid it.

*: technically we could store refresh tokens, but that's (slightly) less secure, and I'm not going to trade real security for perceived security.

Edge and Firefox still use res.com; the Chrome warning should be trustworthy enough, given that chromiumapp.org is owned by Google and isn't quite as obscure as "allizom.org".
We'll see how much hatemail we get.
