+++ browser/background.js
@@ -45,71 +45,6 @@ addListener('ajax', async ({ method, url, headers, data }) => {
 	};
 });
 
-addListener('authFlow', async ({ domain, clientId, scope, interactive }) => {
-	const url = new URL(domain);
-	url.searchParams.set('client_id', clientId);
-	url.searchParams.set('scope', scope);
-	url.searchParams.set('response_type', 'token');
-
-	let responseUrl;
-
-	if (!interactive) {
-		// Try to get an auth token non-interactively using the native API.
-		// The browser will throw an error if it cannot complete without user input.
-		url.searchParams.set('redirect_uri', chrome.identity.getRedirectURL());
-		// chrome.identity is not implemented in Edge, so this throws.
-		// That's okay, because noninteractive auth is expected to fail sometimes.
-		responseUrl = await apiToPromise(chrome.identity.launchWebAuthFlow)({ url: url.href, interactive: false });
-	} else {
-		// Try to get an auth token interactively.
-		// We do not use the native API, because it requires a redirect to a potentially
-		// confusing domain, like chromiumapp.org or allizom.org.
-		// Instead, open a tab which will redirect to the sensible domain redditenhancementsuite.com
-		// and track it with the chrome.tabs API.
-		const redirectUri = 'https://redditenhancementsuite.com/oauth';
-		url.searchParams.set('redirect_uri', redirectUri);
-
-		const { tabs: [{ id }] } = await apiToPromise(chrome.windows.create)({ url: url.href, type: 'popup' });
-
-		responseUrl = await new Promise((resolve, reject) => {
-			function updateListener(tabId, { url }) {
-				if (tabId === id && url && url.startsWith(redirectUri)) {
-					stopListening();
-					// Redirect arrived at the target, send back the URL...
-					resolve(url);
-					// ...and close the tab, since we have no forground script running on it.
-					// It's not feasible to use a foreground script to close the tab anyways,
-					// since Edge blocks it with a ridiculous
-					// "The site you're on is trying to close this window" warning.
-					apiToPromise(chrome.tabs.remove)(id);
-				}
-			}
-
-			function removeListener(tabId) {
-				if (tabId === id) {
-					stopListening();
-					reject(new Error('User cancelled or denied access.'));
-				}
-			}
-
-			function stopListening() {
-				chrome.tabs.onUpdated.removeListener(updateListener);
-				chrome.tabs.onRemoved.removeListener(removeListener);
-			}
-
-			chrome.tabs.onUpdated.addListener(updateListener);
-			chrome.tabs.onRemoved.addListener(removeListener);
-		});
-	}
-
-	const hash = new URL(responseUrl).hash.slice(1);
-	const token = new URLSearchParams(hash).get('access_token');
-
-	if (!token) throw new Error('No token found in response.');
-
-	return token;
-});
-
 addListener('i18n', locale => getLocaleDictionary(locale));
 
 // Chakra bug https://github.com/Microsoft/ChakraCore/issues/2606
