+++ lib/modules/backupAndRestore/providers/GoogleDrive.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import { ajax, launchAuthFlow, Permissions } from '../../../environment';
-import { string } from '../../../utils';
+import { Alert, string } from '../../../utils';
 import Provider from './Provider';
 
 const FILE = 'res-storage.json';
@@ -18,9 +18,12 @@ export default class GoogleDrive extends Provider {
 
 		this.accessToken = await launchAuthFlow({
 			domain: 'https://accounts.google.com/o/oauth2/v2/auth',
-			clientId: '105571180393-8c75sktsum0rnnjl21kco5al3f1lrpsc.apps.googleusercontent.com',
+			clientId: '237880911304-a4040773adrek8uq6dpbg2p7308kf03u.apps.googleusercontent.com',
 			scope: 'https://www.googleapis.com/auth/drive.appdata',
-		});
+		}, message => Alert.open(`
+			<p>RES needs your permission to connect to Google Drive.</p>
+			<p><b>${message}</b></p>
+		`, { cancelable: true }));
 
 		return this;
 	}
